Refactor SatutsPlane
StatusPlane had a global variable, which is generaly considered not a
good parctice. Original status plane was supposed to be something of a
factory, but that plane failed and in the end it became something like a
Singleton and Factori, incorparating worst of both patterns. This commit
refactors StatusView into a separate plane that acts like a regular object.